pyproject.toml
[tool.poetry]
name = "studymate"
version = "0.1.0"
description = ""
authors = ["pecheritsa <pecheritsa.roman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
python-dotenv = "^1.1.0"
pre-commit = "^4.2.0"
psycopg2 = "^2.9.10"
django-autocomplete-light = "^3.12.1"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
